requirements.txt
django-redis
ratelimit
WhiteNoise